pyproject.toml
[project]
name = "hackku26"
version = "0.1.0"
description = "Hack Ku 2026 Project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
